split syntax highlighting modules
